refactorization of trace-analysis
