+++ settings.gradle
@@ -35,7 +35,7 @@ include 'tools:rewrite-log-entries'
 include 'tools:sar'
 include 'tools:trace-analysis'
 include 'tools:trace-analysis-gui'
-include 'tools:trace-analysis-new'
+include 'tools:trace-analysis-old'
 include 'extension-kafka'
 include 'extension-cassandra'
 include 'examples'
