+++ tools/trace-analysis-gui/src/kieker/tools/trace/analysis/gui/AdditionalFiltersStep.java
@@ -28,7 +28,7 @@ import javax.swing.JLabel;
 import javax.swing.JPanel;
 import javax.swing.JTextField;
 
-import kieker.tools.trace.analysis.StringConstants;
+import kieker.tools.common.StringConstants;
 
 /**
  * @author Nils Christian Ehmke
