+++ tools/trace-analysis-gui/src/kieker/tools/trace/analysis/gui/PrintStep.java
@@ -26,7 +26,7 @@ import javax.swing.JLabel;
 import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 
-import kieker.tools.trace.analysis.StringConstants;
+import kieker.tools.common.StringConstants;
 import kieker.tools.trace.analysis.gui.util.AllSelectionBindingItemListener;
 
 /**
