+++ tools/trace-analysis-old/.gitignore
