+++ tools/trace-analysis-old/gradle.properties
@@ -1,9 +1,9 @@
-checkstyleErrorThreshold = 32
+checkstyleErrorThreshold = 0
 checkstyleWarningThreshold = 0
 
 findbugsErrorThreshold = 0
 findbugsWarningThreshold = 2
 
-pmdErrorThreshold = 4
-pmdWarningThreshold = 74
+pmdErrorThreshold = 0
+pmdWarningThreshold = 21
