+++ tools/trace-analysis-old/test/kieker/test/tools/manual/TestAnalysis.java
