+++ tools/trace-analysis/build.gradle
@@ -28,12 +28,6 @@ task uploadArchives {
 	println 'trace-analysis skipped'
 }
 
-task replaceHardCodedVersionNames() {
-	doLast {
-//		regexpReplaceInFiles(project.file("test/kieker/test/tools/manual/TestTCPReader.java"), "kieker-.*-aspectj.jar", "${rootProject.aspectJJar.archiveFileName}")
-	}
-}
-
 eclipse {
 	classpath {
 		file {
@@ -47,5 +41,3 @@ eclipse {
 		}
 	}
 }
-
-compileJava.dependsOn replaceHardCodedVersionNames
