+++ tools/trace-analysis/gradle.properties
@@ -1,9 +1,9 @@
-checkstyleErrorThreshold = 0
+checkstyleErrorThreshold = 32
 checkstyleWarningThreshold = 0
 
 findbugsErrorThreshold = 0
 findbugsWarningThreshold = 2
 
-pmdErrorThreshold = 0
-pmdWarningThreshold = 21
+pmdErrorThreshold = 4
+pmdWarningThreshold = 74
